Consolidated the question code
