--- Assets/Tools/_Scripts/DataPersistence.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ada5af932b5c0da4ca8d32e58697588d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
