--- Assets/Tools/_Scripts/DataPersistence/DataPersistenceManager.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 03400181919d4192a01b2398ccb505ee
-timeCreated: 1669073608
