--- Assets/Tools/_Scripts/DataPersistence/FileDataHandler.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 650b5ea1fd7f417a89533e577b5e283a
-timeCreated: 1669074486
