--- Assets/Tools/_Scripts/DataPersistence/GamePreferences.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: cdd8ed19b45a4fc4be523d4180c5b3cc
-timeCreated: 1669073591
