--- Assets/Tools/_Scripts/Managers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 318217bdf343097499683d41c3fb2617
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
