--- Assets/Tools/_Scripts/MetricCollection.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: abf8cb08750f68b498cf441c47d7e85f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
